% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLBpar.R
\name{BLBpar}
\alias{BLBpar}
\title{Parallelized Bag of Little Bootstraps (BLB) Algorithm}
\usage{
BLBpar(data, beta, s, r, u, xi)
}
\arguments{
\item{data}{A numeric vector representing the input data.}

\item{beta}{A numeric value between 0 and 1, controlling the subset size.}

\item{s}{Number of bootstrap subsets.}

\item{r}{Monte Carlo samples per subset.}

\item{u}{Function to compute point estimates.}

\item{xi}{Function to calculate the statistic (e.g., confidence intervals).}
}
\value{
A numeric vector representing averaged statistics from bootstrap subsets.
}
\description{
Implements a parallelized version of the BLB algorithm.
}
